+++ apps/api_admin/endpoints/token.py
@@ -7,7 +7,7 @@ from apps.api_admin.logics import LoginResolver
 router = APIRouter(route_class=Route, responses=ErrorSchema)
 
 
-@router.post('', response_model=TokenSchema, status_code=201)
+@router.post('', response_model=TokenSchema, status_code=200)
 async def create_token(parser: CreateTokenParser):
     """the api of login and get token"""