+++ extensions/paginate.py
@@ -8,11 +8,11 @@ class Pagination(object):
             self,
             query: QuerySet,
             page: int = PaginateConst.DefaultNum,
-            pagesize: int = PaginateConst.DefaultSize
+            page_size: int = PaginateConst.DefaultSize
     ) -> None:
         self.query = query
         self.page = page
-        self.pagesize = pagesize
+        self.page_size = page_size
 
     def items(self) -> QuerySet[MODEL]:
-        return self.query.offset((self.page - 1) * self.pagesize).limit(self.pagesize)
+        return self.query.offset((self.page - 1) * self.page_size).limit(self.page_size)